requirements.txt
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
google-auth==2.15.0
google-auth-oauthlib==0.8.0
gspread==5.7.2
idna==3.4
numpy==1.23.5
oauthlib==3.2.2
pandas==1.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.13
